Fixes GH-1: Configure whether tests are run on same machine as web server

+++ setenviroment.sh
@@ -1,3 +1,4 @@
+
 base_url=$(cat behat.yml | grep base_url | sed -e "s/[ ]*base_url:[ ]*//")
 read -e -p "Your Base URL (default: $base_url): " new_base_url
 if [ -n "$new_base_url" ]; then
@@ -5,9 +6,23 @@ if [ -n "$new_base_url" ]; then
   rm behat.yml.bak
 fi
 
-drupal_root=$(cat behat.yml | grep drupal_root | sed -e "s/[ ]*drupal_root:[ ]*//")
-read -e -p "Your Drupal root (default: $drupal_root): " new_drupal_root
-if [ -n "$new_drupal_root" ]; then
-  sed -i .bak -e "s!$drupal_root!$new_drupal_root!g" behat.yml
-  rm behat.yml.bak
-fi
+read -p "Run tests from webserver and use Drupal API driver? (y/n):" yesorno
+  case "$yesorno" in
+     y*)
+          drupal_root=$(cat behat.yml | grep drupal_root | sed -e "s/[ ]*drupal_root:[ ]*//")
+          sed -i .bak -e "s/#api_driver/ api_driver/" behat.yml
+          sed -i .bak -e "s/#drupal:/ drupal:/" behat.yml
+          sed -i .bak -e "s/#drupal_root/ drupal_root/" behat.yml
+          read -e -p "Your Drupal root (default: $drupal_root): " new_drupal_root
+            if [ -n "$new_drupal_root" ]; then
+              sed -i .bak -e "s!$drupal_root!$new_drupal_root!g" behat.yml
+              rm behat.yml.bak
+            fi
+          ;;
+     n*)  sed -i .bak -e "s/ api_driver/#api_driver/" behat.yml
+          sed -i .bak -e "s/ drupal:/#drupal:/" behat.yml
+          sed -i .bak -e "s/ drupal_root/#drupal_root/" behat.yml
+         rm behat.yml.bak
+         ;;
+  esac
+cat behat.yml
